move onGridChange to platform-utils.js
