+++ src/platform.js
@@ -1,25 +1,7 @@
-import { getGridColumn } from './ui/grid';
-import debounce from './helpers/debounce';
 import platformUtils from './platform-utils';
 
-function updateGridColumn(callBack) {
-  const orientation = platformUtils.getOrientation(window);
-  const gridColumn = getGridColumn(orientation, platform.platform, platform.deviceType, platform.screenWidth());
-  callBack(gridColumn);
-}
-
 function onGridChange(callBack) {
-  // Triggered once
-  updateGridColumn(callBack);
-
-  if (typeof window !== 'undefined') {
-    window.addEventListener('resize', () => {
-      // Triggered on orientation change
-      debounce(() => {
-        updateGridColumn(callBack);
-      }, 500)();
-    });
-  }
+  platformUtils.onGridChange(window, platform.platform, platform.deviceType, platform.screenWidth, callBack);
 }
 
 const platform = {
